+++ tests/wikidata/test_helpers.py
@@ -3,33 +3,14 @@ from unittest.mock import Mock
 import pytest
 from pytest import MonkeyPatch
 
-from mex.common.types import MergedOrganizationIdentifier, MergedPrimarySourceIdentifier
-from mex.helpers import helpers
-from mex.helpers.helpers import (
-    get_extracted_primary_source_id_by_name,
+from mex.common.types import MergedOrganizationIdentifier
+from mex.extractors.wikidata import helpers
+from mex.extractors.wikidata.helpers import (
     get_wikidata_extracted_organization_id_by_name,
 )
 
 
-@pytest.mark.usefixtures("extracted_primary_sources")
-def test_get_extracted_primary_source_id_by_name(
-    monkeypatch: MonkeyPatch,
-) -> None:
-    """Primary source helper finds "Wikidata" and returns None for nonsense query."""
-    query_wiki = "wikidata"
-    query_nonsense = "this should give None"
-
-    mocked_load = Mock()
-    monkeypatch.setattr(helpers, "load", mocked_load)
-
-    returned = get_extracted_primary_source_id_by_name(query_wiki)
-    mocked_load.assert_called_once()
-
-    assert returned == MergedPrimarySourceIdentifier("djbNGb5fLgYHFyMh3fZE2g")
-    assert get_extracted_primary_source_id_by_name(query_nonsense) is None
-
-
-@pytest.mark.integration
+@pytest.mark.usefixtures("mocked_wikidata")
 def test_get_wikidata_extracted_organization_id_by_name(
     monkeypatch: MonkeyPatch,
 ) -> None:
